Removed the old Zig kernel
The Rust kernel is now the default, and only, version of the kernel.
This means we can finally update everything left to a reasonable version
of Zig, given that 0.7.1 is now practically ancient.
We should also be able to clean up the build system so we don't need as
many dependencies.