Remove old build and docs files
These are pretty old, and don't really apply to the current state of the
project, so keeping them in is just likely to be confusing.
An updated shell.nix should be made at some point, but isn't a top
priority given that most of my current development is done on Windows.
Updated documentation can also be made again once enough functionality
has been implemented for it to be useful.